framework handler register
