+++ aworld/events/util.py
@@ -1,6 +1,5 @@
 # coding: utf-8
 # Copyright (c) 2025 inclusionAI.
-import asyncio
 from typing import Callable, Any
 
 from aworld.core.context.base import Context
@@ -10,7 +9,7 @@ from aworld.events.manager import EventManager
 from aworld.utils.common import sync_exec
 
 
-def subscribe(key: str, category: str = None):
+def subscribe(category: str, key: str = None):
     """Subscribe the special event to handle.
 
     Examples:
@@ -20,15 +19,14 @@ def subscribe(key: str, category: str = None):
         >>>     print("do something")
 
     Args:
+         category: Types of subscription events, the value is `agent` or `tool`, etc.
          key: The index key of the handler.
-         category: Types of subscription events, the value is `agent` or `tool`.
     """
     def decorator(func: Callable[..., Any]) -> Callable[..., Any]:
-        if category is None:
-            sync_exec(eventbus.subscribe, Constants.TOOL, key, func)
-            sync_exec(eventbus.subscribe, Constants.AGENT, key, func)
-        else:
-            sync_exec(eventbus.subscribe, category, key, func)
+        topic = key
+        if not topic:
+            topic = category
+        sync_exec(eventbus.subscribe, category, topic, func)
         return func
 
     return decorator
